chore: replace internal processPipe with a method
