--- lib/process-services/src/lib/pipes/process-services-pipe.module.ts
@@ -1,31 +1,0 @@
-/*!
- * @license
- * Copyright © 2005-2023 Hyland Software, Inc. and its affiliates. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { NgModule } from '@angular/core';
-
-import { ProcessNamePipe } from './process-name.pipe';
-
-@NgModule({
-    declarations: [
-        ProcessNamePipe
-    ],
-    exports: [
-        ProcessNamePipe
-    ]
-})
-export class ProcessServicesPipeModule {
-}
