+++ lib/process-services/src/lib/process-list/process-list.module.ts
@@ -32,7 +32,6 @@ import { ProcessInstanceTasksComponent } from './components/process-instance-tas
 import { ProcessInstanceListComponent } from './components/process-list.component';
 import { StartProcessInstanceComponent } from './components/start-process.component';
 import { FormModule } from '../form/form.module';
-import { ProcessNamePipe } from '../pipes/process-name.pipe';
 
 @NgModule({
     imports: [
@@ -64,10 +63,6 @@ import { ProcessNamePipe } from '../pipes/process-name.pipe';
         ProcessInstanceHeaderComponent,
         ProcessInstanceTasksComponent,
         StartProcessInstanceComponent
-    ],
-    providers: [
-        ProcessNamePipe
     ]
 })
-export class ProcessListModule {
-}
+export class ProcessListModule {}
